Split engine_base from webview.h
