+++ core/include/webview/detail/backends/win32_edge.hh
@@ -37,6 +37,7 @@
 
 #include "../../errors.hh"
 #include "../../types.hh"
+#include "../engine_base.hh"
 #include "../user_script.hh"
 #include "../utility/string.hh"
